+++ dotnet/typeagent/src/knowpro/SemanticRef.cs
@@ -5,8 +5,15 @@ namespace TypeAgent.KnowPro;
 
 public class SemanticRef
 {
+    [JsonPropertyName("semanticRefOrdinal")]
     public SemanticRefOrdinal SemanticRefOrdinal { get; set; }
+
+   // [JsonPropertyName("range")]
     public TextRange Range { get; set; }
+
+   // [JsonPropertyName("knowledgeType")]
     public KnowledgeType KnowledgeType { get; set; }
+
+   // [JsonPropertyName("knowledge")]
     public IKnowledge Knowledge { get; set; }
 }
